+++ include/core/parameters.h
@@ -1567,6 +1567,13 @@ namespace Parameters
    */
   struct MultipleAdaptationParameters
   {
+    /// Error estimator for the variable
+    enum class ErrorEstimator : std::int8_t
+    {
+      kelly,
+      dpg
+    } error_estimator;
+
     // Coarsening fraction
     double coarsening_fraction;
 
@@ -1589,7 +1596,7 @@ namespace Parameters
     {
       none,
       uniform,
-      kelly
+      adaptive
     } type;
 
     /// Fields on which the mesh adaptation can be based
@@ -1601,7 +1608,7 @@ namespace Parameters
     Variable                     vars;
     MultipleAdaptationParameters var_adaptation_param;
 
-    // Decision factor for Kelly refinement (number or fraction)
+    /// Decision factor for adaptive refinement (number or fraction)
     enum class FractionType : std::int8_t
     {
       number,