+++ include/solvers/heat_transfer.h
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: Copyright (c) 2021-2025 The Lethe Authors
+// SPDX-FileCopyrightText: Copyright (c) 2021-2026 The Lethe Authors
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception OR LGPL-2.1-or-later
 
 #ifndef lethe_heat_transfer_h
@@ -235,16 +235,34 @@ public:
   post_mesh_adaptation() override;
 
   /**
-   * @brief Compute the Kelly error estimator for mesh refinement.
+   * @brief Compute the error estimator for mesh refinement.
    *
-   * @param ivar The current element of the map simulation_parameters.mesh_adaptation.variables.
+   * @param[in] ivar The current element of the map
+   * simulation_parameters.mesh_adaptation.variables
    *
-   * @param estimated_error_per_cell The deal.II vector of estimated_error_per_cell.
+   * @param[in,out] estimated_error_per_cell The deal.II vector of
+   * estimated_error_per_cell
    */
   void
-  compute_kelly(const std::pair<const Variable,
-                                Parameters::MultipleAdaptationParameters> &ivar,
-                dealii::Vector<float> &estimated_error_per_cell) override;
+  compute_error_estimate(
+    const std::pair<const Variable, Parameters::MultipleAdaptationParameters>
+                          &ivar,
+    dealii::Vector<float> &estimated_error_per_cell) override;
+
+  /**
+   * @brief Compute the Kelly error estimator on the temperature variable for mesh refinement.
+   * See :
+   * https://www.dealii.org/current/doxygen/deal.II/classKellyErrorEstimator.html
+   * for more information on the Kelly error estimator.
+   *
+   * @param[in,out] estimated_error_per_cell The deal.II vector of
+   * estimated_error_per_cell
+   * @param[in] component_mask The component mask corresponding to the
+   * temperature variable
+   */
+  void
+  compute_kelly(dealii::Vector<float> &estimated_error_per_cell,
+                const ComponentMask   &component_mask);
 
   /**
    * @brief Prepare Heat Transfer to write checkpoint.